Version for updating Github

+++ app/Http/Controllers/AdminPostsController.php
@@ -106,7 +106,10 @@ class AdminPostsController extends Controller
     public function destroy($id)
     {
         $post = Post::findOrFail($id);
-        unlink( public_path() . $post->photo->file );
+
+        if( isset($post->photo->file) ){
+            unlink( public_path() . $post->photo->file );
+        }
         $post->delete();
 
         return redirect('/admin/posts');
